add generic test cases

+++ src/opencl.hpp
@@ -1,7 +1,7 @@
 #ifndef OPENCL_HPP_
 #define OPENCL_HPP_
 
-int run_opencl_cpu(int N);
-int run_opencl_gpu(int N);
+float run_opencl_cpu(int N);
+float run_opencl_gpu(int N);
 
 #endif
